backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
python-jose==3.3.0
python-multipart==0.0.9
httpx==0.27.0
#supabase==2.6.0
python-dotenv==1.1.0
beautifulsoup4==4.12.3
requests==2.32.3
email-validator==2.1.0
pytz==2023.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.29
passlib==1.7.4 
aiohttp==3.9.1
asyncio==3.4.3
selenium==4.31.0
webdriver-manager==4.0.2